requirements.txt
pywin32
pypiwin32
pythoncom
apscheduler
django-apscheduler